+++ NEWS.md
@@ -3,3 +3,5 @@
 * Add `tar_files` as an alternative to `tarchetypes::tar_files` that does not always invalidate the upstream target.
 
 * Add `tar_png` to export a plot as a png.
+
+* Add `tar_csv` to export a dataframe as a csv.